slight refactor of the restapi files
