+++ server/restapis/main.py
@@ -1,55 +1,27 @@
-import tornado.httpserver
-import tornado.ioloop
-import tornado.options
-import tornado.web
-import os.path
-import random
-import string
-from pdfreader import PdfReader
-
-
-class Application(tornado.web.Application):
-    def __init__(self):
-        handlers = [
-            (r"/", IndexHandler),
-            (r"/upload", UploadHandler)
-        ]
-        tornado.web.Application.__init__(self, handlers)
-
-
-class IndexHandler(tornado.web.RequestHandler):
-    def get(self):
-        self.render("tornadoUpload.html")
-
-
-class UploadHandler(tornado.web.RequestHandler):
-    def post(self):
-        files = self.request.files.get('file', [])  # allows a file with any name to be uploaded
-        for file_data in files:
-            original_fname = file_data['filename']
-            upload_path = "uploads/"
-
-            # Check if the directory exists, create it if not
-            if not os.path.exists(upload_path):
-                os.makedirs(upload_path)
-
-            output_file_path = os.path.join(upload_path, original_fname)
-            print("Saving file to:", output_file_path)  # Add a debug statement
-
-            output_file = open(output_file_path, 'wb')
-            output_file.write(file_data['body'])
-            output_file.close()
+from tornado.ioloop import IOLoop
+from tornado.httpserver import HTTPServer
+from tornado.web import Application
+from tornado.log import enable_pretty_logging
+from views import IndexHandler, UploadHandler
 
 
 def make_app():
-    return tornado.web.Application([
+    return Application([
         (r"/", IndexHandler),  # Add a handler for the root URL
         (r"/upload", UploadHandler),
     ])
 
 
+def main(port_number: int):
+    enable_pretty_logging()
+    app = make_app()
+    http_server = HTTPServer(app)
+    http_server.listen(port_number)
+    print(f"Listening on port {port_number}")
+    IOLoop.current().start()
+    
+
 if __name__ == "__main__":
-    app = Application()
-    app.listen(8888)
-    print("Server started.")
-    tornado.ioloop.IOLoop.current().start()
+    print(f"Webserver starting")
+    main(8888)
+    print(f"Webserver ending")
